Combine symlink-related tests
